Move cJSON to ext/, remove debug statements, take advantage of type as a string
